Revert "[DUOS-1411][risk=no] Add Sam Status (#1232)"
This reverts commit 6608e7cc1041d8992e13319ff4cacf52e5df1cd9.
