+++ src/test/java/org/broadinstitute/consent/http/cloudstore/GCSHealthCheckTest.java
@@ -1,16 +1,15 @@
-package org.broadinstitute.consent.http.health;
-
-import com.codahale.metrics.health.HealthCheck;
-import com.google.cloud.storage.Bucket;
-import org.broadinstitute.consent.http.cloudstore.GCSService;
-import org.junit.Before;
-import org.junit.Test;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
+package org.broadinstitute.consent.http.cloudstore;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.when;
+
+import com.codahale.metrics.health.HealthCheck;
+import com.google.cloud.storage.Bucket;
+import org.junit.Before;
+import org.junit.Test;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
 
 public class GCSHealthCheckTest {
 
@@ -24,7 +23,7 @@ public class GCSHealthCheckTest {
 
     @Before
     public void setUpClass() {
-        MockitoAnnotations.openMocks(this);
+        MockitoAnnotations.initMocks(this);
         healthCheck = new GCSHealthCheck(store);
     }
