+++ src/test/java/org/broadinstitute/consent/http/util/OntologyHealthCheckTest.java
@@ -1,16 +1,4 @@
-package org.broadinstitute.consent.http.health;
-
-import com.codahale.metrics.health.HealthCheck;
-import com.google.api.client.http.HttpStatusCodes;
-import org.apache.http.StatusLine;
-import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.entity.StringEntity;
-import org.broadinstitute.consent.http.configurations.ServicesConfiguration;
-import org.broadinstitute.consent.http.util.HttpClientUtil;
-import org.junit.Before;
-import org.junit.Test;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
+package org.broadinstitute.consent.http.util;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -19,7 +7,18 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.when;
 
-public class SamHealthCheckTest {
+import com.codahale.metrics.health.HealthCheck;
+import com.google.api.client.http.HttpStatusCodes;
+import org.apache.http.StatusLine;
+import org.apache.http.client.methods.CloseableHttpResponse;
+import org.broadinstitute.consent.http.configurations.ServicesConfiguration;
+import org.broadinstitute.consent.http.service.ontology.OntologyHealthCheck;
+import org.junit.Before;
+import org.junit.Test;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
+
+public class OntologyHealthCheckTest {
 
   @Mock private HttpClientUtil clientUtil;
 
@@ -29,29 +28,25 @@ public class SamHealthCheckTest {
 
   @Mock private ServicesConfiguration servicesConfiguration;
 
-  private SamHealthCheck healthCheck;
+  private OntologyHealthCheck healthCheck;
 
   @Before
   public void setUp() {
-    MockitoAnnotations.openMocks(this);
+    MockitoAnnotations.initMocks(this);
   }
 
   private void initHealthCheck() {
     try {
-      when(response.getEntity())
-          .thenReturn(
-              new StringEntity(
-                  "{\"ok\":true,\"systems\":{\"GooglePubSub\": {\"ok\": true},\"Database\": {\"ok\": true},\"GoogleGroups\": {\"ok\": true},\"GoogleIam\": {\"ok\": true},\"OpenDJ\": {\"ok\": true}}}"));
       when(clientUtil.getHttpResponse(any())).thenReturn(response);
-      when(servicesConfiguration.getSamUrl()).thenReturn("http://localhost:8000/");
-      healthCheck = new SamHealthCheck(clientUtil, servicesConfiguration);
+      when(servicesConfiguration.getOntologyURL()).thenReturn("http://localhost:8000/");
+      healthCheck = new OntologyHealthCheck(clientUtil, servicesConfiguration);
     } catch (Exception e) {
       fail(e.getMessage());
     }
   }
 
   @Test
-  public void testCheckSuccess() throws Exception {
+  public void testCheckSuccess() {
     when(statusLine.getStatusCode()).thenReturn(HttpStatusCodes.STATUS_CODE_OK);
     when(response.getStatusLine()).thenReturn(statusLine);
     initHealthCheck();
@@ -61,7 +56,7 @@ public class SamHealthCheckTest {
   }
 
   @Test
-  public void testCheckFailure() throws Exception {
+  public void testCheckFailure() {
     when(statusLine.getStatusCode()).thenReturn(HttpStatusCodes.STATUS_CODE_SERVER_ERROR);
     when(response.getStatusLine()).thenReturn(statusLine);
     initHealthCheck();
@@ -71,7 +66,7 @@ public class SamHealthCheckTest {
   }
 
   @Test
-  public void testCheckException() throws Exception {
+  public void testCheckException() {
     doThrow(new RuntimeException()).when(response).getStatusLine();
     initHealthCheck();
